update sierraLinks tests
